+++ rust/tvm-rt/Cargo.toml
@@ -33,7 +33,7 @@ thiserror = "^1.0"
 ndarray = "0.12"
 num-traits = "0.2"
 tvm-sys = { version = "0.1", path = "../tvm-sys/", features = ["bindings"] }
-tvm-macros = { version = "0.1", path = "../macros" }
+tvm-macros = { version = "0.1", path = "../tvm-macros" }
 paste = "0.1"
 mashup = "0.1"
 once_cell = "^1.3.1"
